api/requirements.txt
pandas
numpy==1.26.4
fastapi
uvicorn
qdrant-client
sentence_transformers 
fastembed